cambio restricciones de las vistas a las restricciones

+++ controllers/CinemaController.php
@@ -15,17 +15,21 @@
 
         public function add($name,$adress,$price,$city){
             //tengo que agregar un cine.
-            $cinema = new Cinema($name,$city,$adress,$price);
-            //falta resolver el tema de las salas, por el momento no trabajo con ellas solo se crea un array vacio.
-            $this->cinemaDao->add($cinema);
-            $this->view->admCinema();
+            if(isset($_SESSION['loggedRole']) && $_SESSION['loggedRole'] == 'Admin'){
+                $cinema = new Cinema($name,$city,$adress,$price);
+                //falta resolver el tema de las salas, por el momento no trabajo con ellas solo se crea un array vacio.
+                $this->cinemaDao->add($cinema);
+                $this->view->admCinema();
+            }
 
         }
 
         public function delete($name){
             //borrar el cine con el nombre pasado por parameto
-            $this->cinemaDao->delete($name);
-            $this->view->admCinema();
+            if(isset($_SESSION['loggedRole']) && $_SESSION['loggedRole'] == 'Admin'){
+                $this->cinemaDao->delete($name);
+                $this->view->admCinema();
+            }
         }
 
         public function read($name){
@@ -33,7 +37,9 @@
         }
         public function edit(){
             //editar el valor del atributo seleccionado por el usuario del cine seleccionado
-            $this->cinemaDao->modify($value,$option,$cinemaName);
-            $this->view->admCinema();
+            if(isset($_SESSION['loggedRole']) && $_SESSION['loggedRole'] == 'Admin'){
+                $this->cinemaDao->modify($value,$option,$cinemaName);
+                $this->view->admCinema();
+            }
         }
     }
